requirements.txt
tqdm
opencv-python
pillow
torch
torchvision
numpy
scikit-image
torchsummary
tensorboardX